Separated server code into separate files to improve readability
